+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @opengeni/storage
 
+## 0.2.68
+
+### Patch Changes
+
+- Updated dependencies [5b6d36e]
+- Updated dependencies [6eb0b23]
+  - @opengeni/config@0.11.0
+  - @opengeni/contracts@0.39.0
+
 ## 0.2.67
 
 ### Patch Changes